internal: move SimpleTaskTypes to base-simple-application
