+++ generators/base-simple-application/generator.ts
@@ -22,10 +22,10 @@ import type GeneratorsByNamespace from '../types.js';
 import BaseGenerator from '../base/index.js';
 import { mutateData } from '../../lib/utils/index.js';
 import { GENERATOR_BOOTSTRAP_APPLICATION_BASE } from '../generator-list.js';
-import type { SimpleTaskTypes } from '../../lib/types/application/tasks.js';
 import { getConfigWithDefaults } from '../../lib/jhipster/default-application-options.js';
 import { PRIORITY_NAMES } from '../base/priorities.js';
 import type { GenericTaskGroup } from '../base-core/types.js';
+import type { SimpleTaskTypes } from './tasks.js';
 import { CONTEXT_DATA_APPLICATION_KEY, CONTEXT_DATA_SOURCE_KEY } from './support/index.js';
 import type {
   Application as SimpleApplication,
